Specify behavior when BarcodeDetectorOptions.formats is not present
This change resolves confusion around the |barcodeDetectorOptions|
argument to the BarcodeDetector constructor. As it is an optional
parameter in trailing position it will always be default initialized and
so always present. However, the |formats| member does not have a default
value and so may not be present. This change clarifies the behavior in
these cases.

Fixes #74.