Update links to HTML concepts
This partially fixes #84 by updating the link to the HTML "origin"
concept to be a proper reference to the definition in the HTML
specification. For the image "Broken" state and "origin clean" concepts
however there is no appropriate definition currently exported by the
HTML specification and so these have been replaced with explicit links
rather than autolinks. This is non-ideal but a necessary workaround to
allow this specification to compile while we work on getting the
necessary references exported by HTML.